WORKING VERSION!!!
Now has gradient along (address/max)^2 to provide a more clearly defined gradient.

I haven't made it jar-file compatible or usable for any situations outside of the one right now. There's also a lot of holes in the data, that I'm going to attempt to smooth out. That's what's hopefully going to be fixed in future versions.